hello

+++ src/opr/client/service/IASEServiceAsync.java
@@ -6,8 +6,8 @@ public interface IASEServiceAsync {
 	
 	void setWeightDisplay(String msg, AsyncCallback<Void> callback);
 	void connect(AsyncCallback<Void> callback);
-	void getSWeight(AsyncCallback<Integer> callback);
-	void getSIWeight(AsyncCallback<Integer> callback);
+	void getSWeight(AsyncCallback<Double> callback);
+	void getSIWeight(AsyncCallback<Double> callback);
 	void tara(AsyncCallback<Void> callback);
 
 }
